cal p_value and use black

+++ src/basic_analysis/box/Tokyo/perDay/scatter.py
@@ -149,38 +149,42 @@ for month in ["01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", 
             x_label.append(key)
 
 
-mobile = np.load('/home/is/shuntaro-o/dev/compare_population_and_tweet_number/data/mobile/Tokyostation/Tokyostation_2021.npy')
-tweets = np.load('/home/is/shuntaro-o/dev/compare_population_and_tweet_number/data/twitter/Tokyostation_2021/outlier/Tokyostation_3zi_2021.npy')
-name_key = 'Tokyostation'
+mobile = np.load(
+    "/home/is/shuntaro-o/dev/compare_population_and_tweet_number/data/mobile/Tokyostation/Tokyostation_2021.npy"
+)
+tweets = np.load(
+    "/home/is/shuntaro-o/dev/compare_population_and_tweet_number/data/twitter/Tokyostation_2021/outlier/Tokyostation_3zi_2021.npy"
+)
+name_key = "Tokyostation"
 
 mobile = np.sum(mobile, axis=1)
 tweets = np.sum(tweets, axis=1)
 
 
 list_Week_of_Day = []
-for i in range(0, 365):   
+for i in range(0, 365):
     if i % 7 == 1:
-        list_Week_of_Day.append('Holiday')
+        list_Week_of_Day.append("Holiday")
     if i % 7 == 2:
-        list_Week_of_Day.append('Holiday')
+        list_Week_of_Day.append("Holiday")
     if i % 7 == 0:
-        list_Week_of_Day.append('Workday')
+        list_Week_of_Day.append("Workday")
     if i % 7 == 3:
-        list_Week_of_Day.append('Workday')
+        list_Week_of_Day.append("Workday")
     if i % 7 == 4:
-        list_Week_of_Day.append('Workday')
+        list_Week_of_Day.append("Workday")
     if i % 7 == 5:
-        list_Week_of_Day.append('Workday')
+        list_Week_of_Day.append("Workday")
     if i % 7 == 6:
-        list_Week_of_Day.append('Workday')
-        
+        list_Week_of_Day.append("Workday")
+
 
 df = pd.DataFrame(
-    data=np.stack([mobile, tweets ,list_Week_of_Day]).T,
-    columns=[ "Population", "Tweets_num","Week_of_Day"],
+    data=np.stack([mobile, tweets, list_Week_of_Day]).T,
+    columns=["Population", "Tweets_num", "Week_of_Day"],
 )
-df["Tweets_num"] =df["Tweets_num"].astype(float)
-df["Population"] =df["Population"].astype(float)
+df["Tweets_num"] = df["Tweets_num"].astype(float)
+df["Population"] = df["Population"].astype(float)
 
 
 fig = sns.scatterplot(x="Tweets_num", y="Population", data=df)
@@ -190,10 +194,9 @@ y = tweets.reshape(-1, 1)
 a, b = np.polyfit(X[:, 0], y, 1)
 mi = mutual_info_regression(X, y)
 f_test, _ = f_regression(X, y)
-fig.set(title = "{} Mi={:.2f}".format(name_key, mi[0]))
+fig.set(title="{} Mi={:.2f}".format(name_key, mi[0]))
 print(y.dtype)
 print(mi)
-
 
 
 save_PATH = (
